% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahecc_uq}
\alias{ahecc_uq}
\title{AHECC Classification Codes and Abbreviations}
\format{Data sets \code{ahecc_uq}, \code{ahecc_abb}, \code{ahecc_currencies}
  and \code{ahecc_countries} contain supporting .
  The datasets contain the following variables:
\describe{
  \item{abbreviation}{unit of quantity abbreviation or AHECC abbreviation codes}
  \item{description}{unit of quantity description or AHECC abbreviation description}
  \item{country}{Country name}
  \item{currency_code}{Three-character country currency code}
  \item{currency}{Currency description}
  \item{code}{Two-character country code}
  \item{notes}{AHECC country notes.}
}}
\source{
The AHECC is produced and maintained by the Australian Bureau of Statistics.
}
\usage{
ahecc_uq
}
\description{
Data set containing the Australian Harmonized Export Commodity
  Classification (AHECC) supporting codes and abbreviations
}
\details{
The Australian Harmonized Export Commodity Classification (AHECC) is
  designed for use by exporters, customs brokers and freight forwarders in the
  classification of goods when providing export declarations to the Department
  of Immigration and Border Protection, and to assist users interpret export
  statistics published by the Australian Bureau of Statistics (ABS).

  These data sets provide supporting AHECC Classification Codes and Abbreviations:
  \itemize{
    \item{Unit of Quantity (\code{ahecc_uq})}
    \item{Abbreviations (\code{ahecc_abb})}
    \item{FOB Currency Codes (\code{ahecc_currencies})}
    \item{Countries (\code{ahecc_countries})}
  }
  For further details, see:
  \href{http://www.abs.gov.au/ausstats/abs@.nsf/mf/1233.0}{Australian Harmonized Export Commodity Classification (AHECC)}.
}
\references{
Australian Bureau of Statistics 2016, \emph{Australian Harmonized Export Commodity Classification (AHECC), 2017}, Catalogue no. 1233.0, ABS, Canberra. URL: \url{http://www.abs.gov.au/ausstats/abs@.nsf/mf/1233.0}.
}
\keyword{datasets}
